fix(bench): build storage in default bench scripts; share parseRdfInt with daemon
Round-6 review on #943:
- bench / bench:html / bench:baseline / bench:compare now build
  @origintrail-official/dkg-storage first, so the store-read suite always has
  the worker artefact and can't silently skip the production write-contention
  path under the default bench commands.
- Moved parseRdfInt into the shared metrics-queries module (alongside
  GET_TOTAL_TRIPLES_SPARQL) and use it in both lifecycle.ts and the benchmark,
  so the bench's COUNT parsing can't drift from the daemon's.

cli build + bench tsc clean; bench run unchanged (worker contention intact).
