+++ package.json
@@ -14,15 +14,15 @@
     "test": "turbo test",
     "test:watch": "vitest --config vitest.config.ts",
     "test:coverage": "turbo test:coverage",
-    "bench": "esbench --config esbench.config.mjs",
-    "bench:html": "ESBENCH_HTML=1 ESBENCH_PUBLISH_ASYNC_GET_HTML=1 esbench --config esbench.config.mjs",
+    "bench": "pnpm --filter @origintrail-official/dkg-storage build && esbench --config esbench.config.mjs",
+    "bench:html": "pnpm --filter @origintrail-official/dkg-storage build && ESBENCH_HTML=1 ESBENCH_PUBLISH_ASYNC_GET_HTML=1 esbench --config esbench.config.mjs",
     "bench:analysis": "node --experimental-strip-types bench/analyze-publish-async-get.ts",
     "bench:profile": "node bench/profile-publish-async-get.mjs",
     "bench:store-read": "pnpm --filter @origintrail-official/dkg-storage build && esbench --config esbench.config.mjs --file store-read-latency",
     "bench:swm-large-payload": "pnpm --filter @origintrail-official/dkg benchmark:swm-large-payload --",
     "bench:swm-triple-volume": "pnpm --filter @origintrail-official/dkg benchmark:swm-triple-volume --",
-    "bench:baseline": "ESBENCH_RESULT=bench/results/baseline.json ESBENCH_HTML=1 ESBENCH_HTML_FILE=bench/results/baseline.html esbench --config esbench.config.mjs",
-    "bench:compare": "ESBENCH_DIFF=bench/results/baseline.json ESBENCH_RESULT=bench/results/compare.json ESBENCH_HTML=1 ESBENCH_HTML_FILE=bench/results/compare.html esbench --config esbench.config.mjs",
+    "bench:baseline": "pnpm --filter @origintrail-official/dkg-storage build && ESBENCH_RESULT=bench/results/baseline.json ESBENCH_HTML=1 ESBENCH_HTML_FILE=bench/results/baseline.html esbench --config esbench.config.mjs",
+    "bench:compare": "pnpm --filter @origintrail-official/dkg-storage build && ESBENCH_DIFF=bench/results/baseline.json ESBENCH_RESULT=bench/results/compare.json ESBENCH_HTML=1 ESBENCH_HTML_FILE=bench/results/compare.html esbench --config esbench.config.mjs",
     "lint": "turbo lint",
     "clean": "turbo clean",
     "dev": "turbo dev",
