+++ packages/cli/src/daemon/lifecycle.ts
@@ -14,7 +14,7 @@ import {
   type ServerResponse,
 } from "node:http";
 import { createHash, randomUUID } from "node:crypto";
-import { GET_TOTAL_TRIPLES_SPARQL } from "./metrics-queries.js";
+import { GET_TOTAL_TRIPLES_SPARQL, parseRdfInt } from "./metrics-queries.js";
 import {
   appendFile,
   chmod,
@@ -1797,14 +1797,6 @@ export async function runDaemonInner(
   });
 
   // Extract the plain value from an RDF typed literal like "6"^^<xsd:integer>
-  function parseRdfInt(raw: string | undefined): number {
-    if (!raw) return 0;
-    const m = raw.match(/^"?(\d+)"?\^?\^/);
-    if (m) return parseInt(m[1], 10);
-    const n = parseInt(raw, 10);
-    return isNaN(n) ? 0 : n;
-  }
-
   const metricsSource: MetricsSource = {
     getPeerCount: () =>
       new Set(
