+++ packages/cli/src/daemon/metrics-queries.ts
@@ -1,9 +1,9 @@
-// Canonical SPARQL for the daemon's metrics reads.
+// Canonical SPARQL + result parsing for the daemon's metrics reads.
 //
 // Extracted so cross-package consumers — currently the
 // `bench/store-read-latency` benchmark (regression guard for #939) — can
-// measure the EXACT production read path instead of a duplicated copy that
-// could silently drift.
+// measure the EXACT production read path (both the query AND how its result is
+// parsed) instead of duplicated copies that could silently drift.
 
 /**
  * Total triples across the default graph and all named graphs. Run on the 30s
@@ -13,3 +13,17 @@
  */
 export const GET_TOTAL_TRIPLES_SPARQL =
   'SELECT (COUNT(*) AS ?c) WHERE { { ?s ?p ?o } UNION { GRAPH ?g { ?s ?p ?o } } }';
+
+/**
+ * Parse a SPARQL `COUNT` binding into a number. The value arrives as an RDF
+ * integer literal (e.g. `"1000"^^<…#integer>`) or a bare `1000`; missing or
+ * unparseable values yield 0. Shared by the metrics collector and the
+ * store-read benchmark so result parsing can't drift from the query.
+ */
+export function parseRdfInt(raw: string | undefined): number {
+  if (!raw) return 0;
+  const m = raw.match(/^"?(\d+)"?\^?\^/);
+  if (m) return parseInt(m[1], 10);
+  const n = parseInt(raw, 10);
+  return isNaN(n) ? 0 : n;
+}
